requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.26.0
Werkzeug==2.0.1
gunicorn==20.1.0 